requirements.txt
langchain==0.1.0
langchain-google-genai
langgraph==0.0.30
tavily-python==0.3.0
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
numpy==1.24.3
textblob==0.17.1
python-dotenv==1.0.0
requests==2.31.0
setuptools>=70.0.0
wheel>=0.44.0
